return car and view rentals
ok so these changes should allow the user to see their rentals, and
return them. Sorry this took so long, i got super sick friday and
saturday and on top of that, my laptop refuses to connect to the server
:(. hopefully this isn't coming to late, let me know what i need to
change, since this is likely to be error ridden since i was unable to
debug.